add scripts to close github pull requests
github apparently has no way to opt out of pull requests.

This script is just a copy and paste of sync-github-mirror changed to query
open pull requests and close them. Also there's a systemd timer
to make the script run on all repos once an hour

(might be better to trigger from a github webhook, but that would mean
opening up a webserver, bleh)